--- crates/pgt_hover/src/hovered_item.rs
@@ -1,74 +1,0 @@
-use crate::{contextual_priority::ContextualPriority, to_markdown::ToHoverMarkdown};
-
-/// Mapper type that will be used for filtering and turning data to markdown.
-#[derive(Debug)]
-pub(crate) enum HoverItem<'a> {
-    Table(&'a pgt_schema_cache::Table),
-    Column(&'a pgt_schema_cache::Column),
-    Function(&'a pgt_schema_cache::Function),
-    Role(&'a pgt_schema_cache::Role),
-}
-
-impl<'a> From<&'a pgt_schema_cache::Table> for HoverItem<'a> {
-    fn from(value: &'a pgt_schema_cache::Table) -> Self {
-        HoverItem::Table(value)
-    }
-}
-
-impl<'a> From<&'a pgt_schema_cache::Column> for HoverItem<'a> {
-    fn from(value: &'a pgt_schema_cache::Column) -> Self {
-        HoverItem::Column(value)
-    }
-}
-
-impl<'a> From<&'a pgt_schema_cache::Function> for HoverItem<'a> {
-    fn from(value: &'a pgt_schema_cache::Function) -> Self {
-        HoverItem::Function(value)
-    }
-}
-
-impl<'a> From<&'a pgt_schema_cache::Role> for HoverItem<'a> {
-    fn from(value: &'a pgt_schema_cache::Role) -> Self {
-        HoverItem::Role(value)
-    }
-}
-
-impl ContextualPriority for HoverItem<'_> {
-    fn relevance_score(&self, ctx: &pgt_treesitter::TreesitterContext) -> f32 {
-        match self {
-            HoverItem::Table(table) => table.relevance_score(ctx),
-            HoverItem::Column(column) => column.relevance_score(ctx),
-            HoverItem::Function(function) => function.relevance_score(ctx),
-            HoverItem::Role(role) => role.relevance_score(ctx),
-        }
-    }
-}
-
-impl ToHoverMarkdown for HoverItem<'_> {
-    fn hover_headline<W: std::fmt::Write>(&self, writer: &mut W) -> Result<(), std::fmt::Error> {
-        match self {
-            HoverItem::Table(table) => ToHoverMarkdown::hover_headline(*table, writer),
-            HoverItem::Column(column) => ToHoverMarkdown::hover_headline(*column, writer),
-            HoverItem::Function(function) => ToHoverMarkdown::hover_headline(*function, writer),
-            HoverItem::Role(role) => ToHoverMarkdown::hover_headline(*role, writer),
-        }
-    }
-
-    fn hover_body<W: std::fmt::Write>(&self, writer: &mut W) -> Result<bool, std::fmt::Error> {
-        match self {
-            HoverItem::Table(table) => ToHoverMarkdown::hover_body(*table, writer),
-            HoverItem::Column(column) => ToHoverMarkdown::hover_body(*column, writer),
-            HoverItem::Function(function) => ToHoverMarkdown::hover_body(*function, writer),
-            HoverItem::Role(role) => ToHoverMarkdown::hover_body(*role, writer),
-        }
-    }
-
-    fn hover_footer<W: std::fmt::Write>(&self, writer: &mut W) -> Result<bool, std::fmt::Error> {
-        match self {
-            HoverItem::Table(table) => ToHoverMarkdown::hover_footer(*table, writer),
-            HoverItem::Column(column) => ToHoverMarkdown::hover_footer(*column, writer),
-            HoverItem::Function(function) => ToHoverMarkdown::hover_footer(*function, writer),
-            HoverItem::Role(role) => ToHoverMarkdown::hover_footer(*role, writer),
-        }
-    }
-}
