+++ crates/pgt_hover/src/to_markdown.rs
@@ -1,10 +1,10 @@
 use std::fmt::Write;
 
-use humansize::DECIMAL;
-
 pub(crate) trait ToHoverMarkdown {
     fn hover_headline<W: Write>(&self, writer: &mut W) -> Result<(), std::fmt::Error>;
+
     fn hover_body<W: Write>(&self, writer: &mut W) -> Result<bool, std::fmt::Error>; // returns true if something was written
+
     fn hover_footer<W: Write>(&self, writer: &mut W) -> Result<bool, std::fmt::Error>; // returns true if something was written
 }
 
@@ -28,252 +28,8 @@ pub(crate) fn format_hover_markdown<T: ToHoverMarkdown>(
     Ok(markdown)
 }
 
-impl ToHoverMarkdown for pgt_schema_cache::Table {
-    fn hover_headline<W: Write>(&self, writer: &mut W) -> Result<(), std::fmt::Error> {
-        write!(writer, "`{}.{}`", self.schema, self.name)?;
-
-        let table_kind = match self.table_kind {
-            pgt_schema_cache::TableKind::View => " (View)",
-            pgt_schema_cache::TableKind::MaterializedView => " (M.View)",
-            pgt_schema_cache::TableKind::Partitioned => " (Partitioned)",
-            pgt_schema_cache::TableKind::Ordinary => "",
-        };
-
-        write!(writer, "{}", table_kind)?;
-
-        let locked_txt = if self.rls_enabled {
-            " - 🔒 RLS enabled"
-        } else {
-            " - 🔓 RLS disabled"
-        };
-
-        write!(writer, "{}", locked_txt)?;
-
-        Ok(())
-    }
-
-    fn hover_body<W: Write>(&self, writer: &mut W) -> Result<bool, std::fmt::Error> {
-        if let Some(comment) = &self.comment {
-            write!(writer, "{}", comment)?;
-            Ok(true)
-        } else {
-            Ok(false)
-        }
-    }
-
-    fn hover_footer<W: Write>(&self, writer: &mut W) -> Result<bool, std::fmt::Error> {
-        write!(
-            writer,
-            "~{} rows, ~{} dead rows, {}",
-            self.live_rows_estimate,
-            self.dead_rows_estimate,
-            humansize::format_size(self.bytes as u64, DECIMAL)
-        )?;
-        Ok(true)
-    }
-}
-
-impl ToHoverMarkdown for pgt_schema_cache::Column {
-    fn hover_headline<W: Write>(&self, writer: &mut W) -> Result<(), std::fmt::Error> {
-        write!(
-            writer,
-            "`{}.{}.{}`",
-            self.schema_name, self.table_name, self.name
-        )
-    }
-
-    fn hover_body<W: Write>(&self, writer: &mut W) -> Result<bool, std::fmt::Error> {
-        if let Some(tname) = self.type_name.as_ref() {
-            write!(writer, "`{}", tname)?;
-            if let Some(l) = self.varchar_length {
-                write!(writer, "({})", l)?;
-            }
-            write!(writer, "`")?;
-        } else {
-            write!(writer, "typeid: `{}`", self.type_id)?;
-        }
-
-        if self.is_primary_key {
-            write!(writer, " - 🔑 primary key")?;
-        } else if self.is_unique {
-            write!(writer, " - unique")?;
-        }
-
-        if self.is_nullable {
-            write!(writer, " - nullable")?;
-        } else {
-            write!(writer, " - not null")?;
-        }
-
-        if let Some(comment) = &self.comment {
-            write!(writer, "  \n{}", comment)?;
-        }
-
-        Ok(true)
-    }
-
-    fn hover_footer<W: Write>(&self, writer: &mut W) -> Result<bool, std::fmt::Error> {
-        if let Some(default) = &self.default_expr {
-            write!(writer, "Default: `{}`", default)?;
-            Ok(true)
-        } else {
-            Ok(false)
-        }
-    }
-}
-
-impl ToHoverMarkdown for pgt_schema_cache::Function {
-    fn hover_headline<W: Write>(&self, writer: &mut W) -> Result<(), std::fmt::Error> {
-        write!(writer, "`{}.{}", self.schema, self.name)?;
-
-        if let Some(args) = &self.argument_types {
-            write!(writer, "({})", args)?;
-        } else {
-            write!(writer, "()")?;
-        }
-
-        write!(
-            writer,
-            " → {}`",
-            self.return_type.as_ref().unwrap_or(&"void".to_string())
-        )?;
-
-        Ok(())
-    }
-
-    fn hover_body<W: Write>(&self, writer: &mut W) -> Result<bool, std::fmt::Error> {
-        let kind_text = match self.kind {
-            pgt_schema_cache::ProcKind::Function => "Function",
-            pgt_schema_cache::ProcKind::Procedure => "Procedure",
-            pgt_schema_cache::ProcKind::Aggregate => "Aggregate",
-            pgt_schema_cache::ProcKind::Window => "Window",
-        };
-
-        write!(writer, "{}", kind_text)?;
-
-        let behavior_text = match self.behavior {
-            pgt_schema_cache::Behavior::Immutable => " - Immutable",
-            pgt_schema_cache::Behavior::Stable => " - Stable",
-            pgt_schema_cache::Behavior::Volatile => "",
-        };
-
-        write!(writer, "{}", behavior_text)?;
-
-        if self.security_definer {
-            write!(writer, " - Security DEFINER")?;
-        } else {
-            write!(writer, " - Security INVOKER")?;
-        }
-
-        Ok(true)
-    }
-
-    fn hover_footer<W: Write>(&self, writer: &mut W) -> Result<bool, std::fmt::Error> {
-        if let Some(def) = self.definition.as_ref() {
-            /*
-             * We don't want to show 250 lines of functions to the user.
-             * If we have more than 30 lines, we'll only show the signature.
-             */
-            if def.lines().count() > 30 {
-                let without_boilerplate: String = def
-                    .split_ascii_whitespace()
-                    .skip_while(|elem| {
-                        ["create", "or", "replace", "function"]
-                            .contains(&elem.to_ascii_lowercase().as_str())
-                    })
-                    .collect::<Vec<&str>>()
-                    .join(" ");
-
-                for char in without_boilerplate.chars() {
-                    match char {
-                        '(' => {
-                            write!(writer, "(\n  ")?;
-                        }
-
-                        ')' => {
-                            write!(writer, "\n)\n")?;
-                            break;
-                        }
-
-                        ',' => {
-                            // one space already present
-                            write!(writer, ",\n ")?;
-                        }
-
-                        _ => {
-                            write!(writer, "{}", char)?;
-                        }
-                    }
-                }
-            } else {
-                write!(writer, "```\n{}\n```", def)?;
-            }
-
-            Ok(true)
-        } else {
-            Ok(false)
-        }
-    }
-}
-
-fn markdown_newline<W: Write>(writer: &mut W) -> Result<(), std::fmt::Error> {
+pub(crate) fn markdown_newline<W: Write>(writer: &mut W) -> Result<(), std::fmt::Error> {
     write!(writer, "  ")?;
     writeln!(writer)?;
     Ok(())
-}
-
-impl ToHoverMarkdown for pgt_schema_cache::Role {
-    fn hover_headline<W: Write>(&self, writer: &mut W) -> Result<(), std::fmt::Error> {
-        write!(writer, "`{}`", self.name)?;
-
-        if self.is_super_user {
-            write!(writer, " - 🔑 superuser")?;
-        }
-
-        if self.can_login {
-            write!(writer, " - 👤 login")?;
-        }
-
-        if self.can_create_db {
-            write!(writer, " - 🏗️ create DB")?;
-        }
-
-        if self.can_bypass_rls {
-            write!(writer, " - 🛡️ bypass RLS")?;
-        }
-
-        Ok(())
-    }
-
-    fn hover_body<W: Write>(&self, writer: &mut W) -> Result<bool, std::fmt::Error> {
-        let role_type = if self.is_super_user {
-            "Superuser role"
-        } else if self.can_login {
-            "Login role"
-        } else {
-            "Group role"
-        };
-
-        write!(writer, "{}", role_type)?;
-
-        let mut capabilities = Vec::new();
-
-        if self.can_create_db && !self.is_super_user {
-            capabilities.push("create databases");
-        }
-
-        if self.can_bypass_rls && !self.is_super_user {
-            capabilities.push("bypass row-level security");
-        }
-
-        if !capabilities.is_empty() {
-            write!(writer, " with permission to {}", capabilities.join(", "))?;
-        }
-
-        Ok(true)
-    }
-
-    fn hover_footer<W: Write>(&self, _writer: &mut W) -> Result<bool, std::fmt::Error> {
-        Ok(false)
-    }
 }
